java commit

+++ src/basicJava/Irk_06_Cal_Percentage.java
@@ -4,20 +4,21 @@ public class Irk_06_Cal_Percentage {
     public static void main(String[] args)throws Exception {
 
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        int phy,math,chem,eng,hindi;
+        float phy,math,chem,eng,hindi;
         System.out.println("Enter five subject :");
-        phy = Integer.parseInt(br.readLine());
+//        phy = Integer.parseInt(br.readLine());
+        phy = Float.parseFloat(br.readLine());
         System.out.println("Physics Marks = "+phy);
-        math = Integer.parseInt(br.readLine());
+        math = Float.parseFloat(br.readLine());
         System.out.println("Math Marks = "+math);
-        chem = Integer.parseInt(br.readLine());
+        chem = Float.parseFloat(br.readLine());
         System.out.println("Chemistry Marks = "+chem);
-        eng = Integer.parseInt(br.readLine());
+        eng = Float.parseFloat(br.readLine());
         System.out.println("English Marks = "+eng);
-        hindi = Integer.parseInt(br.readLine());
+        hindi = Float.parseFloat(br.readLine());
         System.out.println("Hindi Marks = "+hindi);
 
-        int total = ((phy+math+chem+eng+hindi)/5);
+        float total = ((phy+math+chem+eng+hindi)/5.0f);
         System.out.println("Total percentage = "+total);
 
     }
